[기능추가] param 동작 추가
